+++ source/include/Ostap/LorentzVectorWithError.h
@@ -330,6 +330,22 @@ namespace Ostap
     ( const LorentzVectorWithError::Vector&  b , 
       const LorentzVectorWithError&          a ) { return a.chi2  ( b ) ; }    
     // ========================================================================
+    /** Get symmetrized Kullback-Leibler divergency for two objects 
+     *  @see https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
+     *  @see Ostap::Math::kullback_leibler 
+     */
+    double kullback_leibler 
+    ( const LorentzVectorWithError& a , 
+      const LorentzVectorWithError& b ) ;
+    // ========================================================================
+    /** Get asymmetric Kullback-Leibler divergency for two objects 
+     *  @see https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
+     *  @see Ostap::Math::asymmetric_kullback_leibler 
+     */
+    double asymmetric_kullback_leibler 
+    ( const LorentzVectorWithError& a , 
+      const LorentzVectorWithError& b ) ;
+    // ======================================================================
     namespace Kinematics 
     {
       // ======================================================================
@@ -536,7 +552,7 @@ namespace Ostap
   // ==========================================================================
 } //                                                     end of namespace Ostap
 // ============================================================================
-// The END 
+//                                                                      The END 
 // ============================================================================
 #endif // OSTAP_LORENTZVECTORWITHERROR_H
 // ============================================================================