+++ source/include/Ostap/Vector3DWithError.h
@@ -131,11 +131,13 @@ namespace Ostap
     public:
       // ======================================================================
       /// get generic vector 
-      void    asVector ( Vector&  data ) const ;
+      void    asVector  ( Vector&  data ) const ;
       /// get generic vector 
-      void    asVector ( VectorE& data ) const ;
+      void    asVector  ( VectorE& data ) const ;
       /// convert to generic vector with errors:
-      VectorE asVector () const ;
+      VectorE asVector  () const ;
+      /// convert to generic vector
+      Vector  asVector3 () const ;
       // ======================================================================
     public:
       // ======================================================================
@@ -254,11 +256,27 @@ namespace Ostap
     ( const Vector3DWithError& v1 , 
       const Vector3DWithError& v2 ) { return v1.mean ( v2 ) ; }
     // ========================================================================
+    /** Get symmetrized Kullback-Leibler divergency for two objects 
+     *  @see https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
+     *  @see Ostap::Math::kullback_leibler 
+     */
+    double kullback_leibler 
+    ( const Vector3DWithError& a , 
+      const Vector3DWithError& b ) ;
+    // ========================================================================
+    /** Get asymmetric Kullback-Leibler divergency for two objects 
+     *  @see https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
+     *  @see Ostap::Math::asymmetric_kullback_leibler 
+     */
+    double asymmetric_kullback_leibler 
+    ( const Vector3DWithError& a , 
+      const Vector3DWithError& b ) ;
+    // ========================================================================
   } //                                             end of namespace Ostap::Math 
   // ==========================================================================
 } //                                                    end of namespace  Ostap
 // ============================================================================
-// The END 
+//                                                                      The END 
 // ============================================================================
 #endif // LHCBMATH_POINTWITHERROR_H
 // ============================================================================