+++ source/src/Point3DWithError.cpp
@@ -316,6 +316,17 @@ Ostap::Math::Point3DWithError::asVector () const
   return data ;
 }
 // ============================================================================
+Ostap::Math::Point3DWithError::Vector
+Ostap::Math::Point3DWithError::asVector3 () const
+{
+  Ostap::Math::Point3DWithError::Vector data ;
+  Ostap::Math::geo2LA ( point() , data ) ;
+  return data ;
+}
+// ============================================================================
+
+
+
 
 // ============================================================================
 void Ostap::Math::Point3DWithError::setValue
@@ -378,9 +389,36 @@ Ostap::Math::Point3DWithError::mean
 ( const Ostap::Math::Point3DWithError::VectorE& right ) const 
 { return asVector ().mean ( right            ) ; }
 // ============================================================================
-
+/* Get symmetrized Kullback-Leibler divergency for two objects 
+ *  @see https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
+ *  @see Ostap::Math::kullback_leibler 
+ */
+// ============================================================================
+double Ostap::Math::kullback_leibler 
+( const Ostap::Math::Point3DWithError& a , 
+  const Ostap::Math::Point3DWithError& b ) 
+{
+  return Ostap::Math::kullback_leibler 
+    ( a.asVector3() , a.covariance () , 
+      b.asVector3() , b.covariance () ) ;
+}
+// ============================================================================
+/* Get asymmetric Kullback-Leibler divergency for two objects 
+ *  @see https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
+ *  @see Ostap::Math::asymmetric_kullback_leibler 
+ */
+// ============================================================================
+double Ostap::Math::asymmetric_kullback_leibler 
+( const Ostap::Math::Point3DWithError& a , 
+  const Ostap::Math::Point3DWithError& b ) 
+{
+  return Ostap::Math::asymmetric_kullback_leibler 
+    ( a.asVector3() , a.covariance () , 
+      b.asVector3() , b.covariance () ) ;
+}
+// ============================================================================
 
 // ============================================================================
-// The END
+//                                                                      The END
 // ============================================================================